# Conflicts:
#	Session/Conversations/ConversationVC+Interaction.swift
#	Session/Conversations/ConversationVC.swift
#	Session/Conversations/Settings/ThreadSettingsViewModel.swift
#	Session/Shared/Views/SessionCell+AccessoryView.swift
#	SessionMessagingKit/Database/Models/SessionThread.swift
#	SessionMessagingKit/Utilities/ProfilePictureView+Convenience.swift
#	SessionUIKit/Types/IconSize.swift

+++ SessionMessagingKit/Database/Models/DisappearingMessageConfiguration.swift
@@ -196,7 +196,7 @@ public extension DisappearingMessagesConfiguration {
     }
     
     var durationString: String {
-        floor(durationSeconds).formatted(format: .long)
+        floor(durationSeconds).formatted(format: .short)
     }
     
     func messageInfoString(
